move api-requests into seperate api-file
